feat(game): parameterize chunk generator for in memory storage
