+++ game/src/main/kotlin/com/github/bedrockecs/game/db/world/WorldDBStorage.kt
@@ -2,7 +2,8 @@ package com.github.bedrockecs.game.db.world
 
 import com.github.bedrockecs.game.data.ChunkPosition
 import com.github.bedrockecs.game.db.world.serial.SerialChunk
-import com.github.bedrockecs.game.zimpl.db.world.InMemoryEmptyWorldDBStorage
+import com.github.bedrockecs.game.zimpl.db.world.EmptyChunkGenerator
+import com.github.bedrockecs.game.zimpl.db.world.InMemoryWorldDBStorage
 import java.util.concurrent.CompletableFuture
 
 /**
@@ -11,13 +12,21 @@ import java.util.concurrent.CompletableFuture
 interface WorldDBStorage {
     companion object {
         /**
+         * mock storage that keeps everything in memory
+         */
+        fun inMemory(
+            heightRange: Pair<Int, Int>,
+            generator: (ChunkPosition) -> SerialChunk
+        ): WorldDBStorage {
+            return InMemoryWorldDBStorage(heightRange, generator)
+        }
+
+        /**
          * mock storage that keeps everything in memory and generates empty chunk
          */
-        fun inMemoryEmpty(
-            heightRange: Pair<Int, Int>,
-            serial: WorldDBSerial
-        ): WorldDBStorage {
-            return InMemoryEmptyWorldDBStorage(heightRange, serial)
+        fun inMemoryEmpty(heightRange: Pair<Int, Int>, serial: WorldDBSerial): WorldDBStorage {
+            val generator = EmptyChunkGenerator(heightRange, serial)
+            return inMemory(heightRange, generator::generate)
         }
     }
