+++ game/src/test/kotlin/com/github/bedrockecs/game/zimpl/db/world/InMemoryWorldDBStorageTest.kt
@@ -14,7 +14,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
-internal class InMemoryEmptyWorldDBStorageTest {
+internal class InMemoryWorldDBStorageTest {
 
     val testSerial = WorldDBSerial.createAutoAssigning(TestAirBlockType.primary)
